apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
recipe-scrapers==15.4.0
python-multipart==0.0.6
pydantic==2.5.2
httpx>=0.24.0,<0.25.0
supabase==2.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
ingredient-parser-nlp
